feat(api): restore live status endpoint
